Statepoint: Fix handling of Far Immediate calls
gc.statepoint intrinsics with a far immediate target was lowered incorrectly as a pc-rel32 call.

For example, for
 %safepoint_token = call i32 (i64, i32, void ()*, i32, i32, ...) @llvm.experimental.gc.statepoint.p0f_isVoidf(i64 0, i32 0, void ()* inttoptr (i64 140727162896504 to void ()*), i32 0, i32 0, i32 0, i32 0)

the CALL instruction generated was:

callq 140727162896504

This change fixes the problem, so that an indirect call via a scratch register is correctly generated:

movabsq $140727162896504, %rax
 callq *%rax

In lowerCallFromStatepoint(), the callee-target was modified and represented as a "TargetConstant" node, rather than a "Constant" node.
 Undoing this modification enabled LowerCall() to generate the correct CALL instruction.